Desktop/My Projects/webappkaitraining/world_class_app/backend/requirements.txt
fastapi
uvicorn
google-generativeai
supabase
python-multipart
pydantic-settings
requests
pydantic
python-dotenv
